Revert "08 add comment"
